Tentative fix for spellslinger voidslip breaking timers.
Wait 7s after player is out of combat before stopping timers.

This is because when a Spellslinger enters the void, all player units
disappear and the slinger goes out of combat, so we have no reliable way
to tell this situation apart from a regular "going out of combat"
situation.